# docs/02-architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: Forge Platform API
  version: 1.0.0
  description: >-
    REST contract consumed by the Forge SPA (catalog, approval queue, trace
    viewer). Consistent with the data model (data-model.md) and the agent DNA
    contract (dna-schema.json). Governance is structural: publishing is
    eval-gated (409 when the gate has not passed), approvals expire server-side
    and there is no extend/auto-approve operation, and a role header
    demonstrates segregation of duties (NFR-5). This is a demonstration
    contract, not a production auth design.
servers:
  - url: /api/v1
...
paths:
...
  /runs:
    post:
      tags: [Runs]
      summary: Start a run of a published agent version
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartRun' }
      responses:
        '202':
          description: Run accepted and started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            The version cannot run. A draft answers code
            agent_version_not_published. A suspended version answers code
            agent_suspended with the governance reason code and the recorded
            suspension in details — and the refusal itself is appended as a
            governance.run_refused event, so a contained agent's refusals are
            visible in the log, not just absent from it (FR-G4).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: agent_suspended
                message: >-
                  This agent is suspended, so the platform refused to start the
                  run. A suspension never lifts itself: an authorised person
                  must resume the agent, and both the suspension and the resume
                  are on the record.
                details:
                  reason_code: agent_suspended
                  status: suspended
                  suspension:
                    trigger: circuit_breaker
                    detail: >-
                      circuit breaker tripped: 4 of 6 finished run(s) in the
                      last 3600s faulted (rate 0.667, threshold 0.500)
  /runs/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      summary: Get run status and summary
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /runs/{runId}/trace:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      summary: Get the full run trace (steps, tool calls, decisions)
      description: >-
        A projection of the append-only events for this run (ADR-008). Every
        model call, tool invocation, rule applied, decision, and **governance
        refusal** is present and ordered, so the run is reconstructable (FR-G1).
        A run the platform stopped carries exactly one step of kind
        `governance`, holding the machine-readable reason code and the
        explanation that goes with it.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Ordered trace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunTrace' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    RoleHeader:
      name: X-Forge-Role
      in: header
      required: true
      description: >-
        Demonstration role for segregation of duties (NFR-5). configurator
        holds read, agent.configure, agent.publish, agent.suspend and
        run.start; approver holds read and approval.decide; viewer holds read
        only; admin holds read, agent.suspend and agent.resume — the
        containment operator, deliberately not a superuser. No role holds both
        agent.configure/agent.publish and approval.decide, and none holds both
        agent.configure/agent.publish and agent.resume: the person who decides
        what an agent may do is never the person who approves what it proposes,
        and whoever built or shipped an agent can never override the breaker
        that contained it. The build refuses to start if either pairing ever
        stops being true. A role lacking the permission for an operation gets
        403, and the refusal is recorded as a governance.permission_denied
        event. Not authentication.
      schema:
        type: string
        enum: [configurator, approver, viewer, admin]
...
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed request or DNA that fails schema validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The X-Forge-Role is not permitted to perform this operation (NFR-5)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine-readable error code }
        message: { type: string, description: Human-readable summary }
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. validation errors)
...
    StartRun:
      type: object
      required: [agent_id, version, input]
      properties:
        agent_id: { type: string, format: uuid }
        version: { type: string, description: semver of a published version }
        input:
          type: object
          additionalProperties: true
          description: Trigger payload (e.g. the invoice reference to process)
    Run:
      type: object
      required: [id, tenant_id, agent_version_id, status]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        agent_version_id: { type: string, format: uuid }
        status:
          type: string
          enum: [running, awaiting_approval, completed, escalated, canceled, error]
        trigger: { type: string }
        total_tokens: { type: integer }
        total_cost_usd:
          type: [string, 'null']
          description: >-
            Exact decimal, serialised as a string. Money is never carried as a
            JSON number here: an audit figure rounded through a float is not an
            audit figure. The same convention applies to cost_usd inside a
            model.called event payload.
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
    RunTrace:
      type: object
      required: [run_id, steps]
      properties:
        run_id: { type: string, format: uuid }
        steps:
          type: array
          items: { $ref: '#/components/schemas/RunStep' }
        events:
          type: array
          description: >-
            The raw append-only events the steps were projected from, in
            append order, including lifecycle events that are not steps
            (run.started, run.completed, run.escalated, run.awaiting_approval,
            run.failed). Served alongside the projection so a reviewer can check
            the trace against its source rather than trusting it (ADR-008).
          items: { $ref: '#/components/schemas/RunEvent' }
    RunEvent:
      type: object
      required: [event_id, type, actor, occurred_at, payload]
      properties:
        event_id:
          type: integer
          format: int64
          description: Monotonic identity; ordering is part of the audit contract
        type:
          type: string
          description: >-
            run.started, model.called, tool.called, decision.made,
            governance.blocked, approval.pending, approval.granted,
            approval.rejected, approval.expired, run.completed, run.escalated,
            run.awaiting_approval, run.canceled, run.failed. The approval events
            carry the deciding actor and the timestamp (FR-E4); approval.expired
            is recorded against `system`, because expiry is the absence of a
            decision rather than one. Outside a run,
            governance.permission_denied records an operation refused to a role
            that lacked its permission; governance.run_refused records a start
            refused because the version is suspended (no run exists); and
            version.suspended / version.resumed record the containment
            lifecycle itself — who or what stopped an agent, and who put it
            back (FR-G4).
        actor: { type: string, description: system, seed-script, or role:<role> }
        occurred_at: { type: string, format: date-time }
        payload:
          type: object
          additionalProperties: true
          description: Typed by event type
    RunStep:
      type: object
      required: [step_no, kind]
      properties:
        step_no: { type: integer }
        kind:
          type: string
          enum: [reason, tool, decision, governance, approval]
          description: >-
            Who acted at this position in the run. `reason`, `tool` and
            `decision` are the agent; `governance` is the platform refusing; and
            `approval` is a **person** — the action the run parked, and the
            approve / reject / expire that answered it (FR-E4).
        model_call:
          type: object
          additionalProperties: true
          description: model id, tokens, cost when kind is reason
        decision:
          type: object
          description: Final action and rule citations when kind is decision
          properties:
            action:
              type: string
              enum: [auto_approve, escalate, block_escalate, priority_queue]
            citations:
              type: array
              minItems: 1
              items:
                type: string
                description: >-
                  Rule ID (R-001) or document citation
                  (AP-Policy-2023.pdf#approval-thresholds)
              description: >-
                What the decision applied, under one citation model (FR-D4):
                governed rule IDs and, for any claim drawn from retrieved
                knowledge, the document#section citations retrieval returned.
                Non-empty by contract (R-092): a decision without citations is
                rejected by the runtime, not merely logged.
            reasoning:
              type: string
              description: Why the action follows from those rules
            confidence:
              type: number
              minimum: 0
              maximum: 1
              description: >-
                How confident the agent is in this decision. Required: R-091
                makes low confidence a fail-closed condition, and a floor cannot
                be enforced against a number the agent was free to omit. Below
                the agent's declared guardrails.min_decision_confidence the
                runtime overrides the action and escalates, recording a
                governance step with reason_code low_confidence.
            output:
              type: object
              additionalProperties: true
              description: >-
                Optional agent-specific structured result, for agents whose job
                is extraction rather than adjudication — the normalised invoice
                from the intake agent, for example. Omitted (not null) when the
                agent produced none. The governed part of a decision is always
                the action and its citations; this sits beside them.
        tool_invocation: { $ref: '#/components/schemas/ToolInvocation' }
        governance: { $ref: '#/components/schemas/GovernanceBlock' }
        approval: { $ref: '#/components/schemas/ApprovalStep' }
        created_at: { type: string, format: date-time }
    ApprovalStep:
      type: object
      description: >-
        One state of one approval, present when kind is `approval`. Carries the
        action being decided — tool ref and the exact arguments — because that is
        the scope of the approval, so the log answers "what was authorised"
        without a join and a later edit to the invocation row could not rewrite
        what somebody approved.
      required: [approval_id, status, tool_ref, expires_at]
      properties:
        approval_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, granted, rejected, expired] }
        tool_ref: { type: string }
        args: { type: object, additionalProperties: true }
        expires_at: { type: string, format: date-time }
        decided_by: { type: [string, 'null'] }
        decided_at: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'] }
        reason_code:
          oneOf:
            - $ref: '#/components/schemas/ReasonCode'
            - type: 'null'
          description: >-
            `approval_rejected` or `approval_expired` when the approval ended the
            run. Null while pending and on a grant, which ends nothing.
    ToolInvocation:
      type: object
      required: [id, tool_ref, autonomy, status]
      properties:
        id: { type: string, format: uuid }
        tool_ref: { type: string, description: slug@semver }
        autonomy: { type: string, enum: [autonomous, requires_approval, forbidden] }
        args: { type: object, additionalProperties: true }
        result: { type: object, additionalProperties: true }
        status:
          type: string
          enum: [validated, executed, blocked, denied]
          description: >-
            executed — allowed and run. validated — checked and parked: the DNA
            grants this tool only with a human approval, so the call did not run
            and the run stopped in awaiting_approval (FR-E2). blocked — refused
            by policy or validation. denied — the DNA forbids the tool outright.
        error:
          type: [string, 'null']
          description: >-
            Why the gateway refused, when status is blocked or denied — unknown
            tool, missing grant, forbidden tool, invalid arguments, or a config
            the tool cannot honour (FR-C5). Null for an executed or a validated
            (parked) invocation.
        reason_code:
          oneOf:
            - $ref: '#/components/schemas/ReasonCode'
            - type: 'null'
          description: >-
            The governance code the gateway assigned to this refusal, or null
            when the call executed. The same code appears on the governance step
            that follows, so the refusal and the stop it caused are one story.
        approval_id:
          type: [string, 'null']
          format: uuid
          description: >-
            Set when this call ran only because a person released it. Null for
            an autonomous execution — an action a human signed for must never
            read the same as one the agent took on its own (FR-E4).
        released_by:
          type: [string, 'null']
          description: Who granted the approval that let this call run
    ReasonCode:
      type: string
      description: >-
        Why the platform stopped, blocked, or refused. Defined once in
        app/governance.py and used unchanged by the runtime, the audit log, this
        API, and the UI, so a refusal reads the same everywhere.
      enum:
        - tool_unknown
        - permission_denied
        - args_invalid
        - tool_config_invalid
        - tool_failed
        - approval_required
        - approval_rejected
        - approval_expired
        - no_rule_match
        - knowledge_conflict
        - low_confidence
        - invalid_output
        - step_limit
        - timeout
        - budget_exceeded
        - daily_budget_exceeded
        - agent_suspended
        - provider_unavailable
        - unsupported_definition
        - agent_decision
    GovernanceBlock:
      type: object
      description: >-
        One platform refusal, present when kind is governance. A run carries one
        per stop, and it always says why (FR-C5, FR-G2). Normally that is exactly
        one; a run that paused for a human approval records the refusal that
        paused it (`approval_required`) and then, if it stops again, the one that
        ended it — a rejection, an expiry, or whatever the resumed loop hit.
      required: [reason_code, explanation, terminal_status]
      properties:
        reason_code: { $ref: '#/components/schemas/ReasonCode' }
        explanation:
          type: string
          description: >-
            The plain-language sentence recorded alongside the code, written for
            a reader who has never seen the codebase. Served rather than left to
            each client so the words a reviewer reads are the words the platform
            recorded when it acted.
        detail:
          type: [string, 'null']
          description: >-
            The specific circumstance — which tool, which ceiling, which value.
        terminal_status:
          type: string
          enum: [escalated, awaiting_approval, canceled, error]
          description: >-
            The state the run ended in as a result of this refusal.
            `awaiting_approval` is a pause rather than an ending — the run
            resumes on a grant; `canceled` is where a rejected or expired
            approval leaves it, with nothing executed.
...